Added fileCopy helper function and unit test
